# proto/openapi/argosy.yaml
openapi: 3.0.3
info:
  title: Argosy API
  version: 0.1.0
  description: |
    Self-hosted media streaming server. This spec is the single source of truth
    for the HTTP contract; the Go server interface and the TypeScript client are
    generated from it (`make generate`), and CI fails on drift.
servers:
  - url: /
...
paths:
...
  /api/v1/libraries:
    get:
      tags: [library]
      operationId: listLibraries
      summary: List the account's libraries
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Library"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /api/v1/libraries/{libraryId}/movies:
    get:
      tags: [library]
      operationId: listMovies
      summary: Browse movies in a library
      parameters:
        - { name: libraryId, in: path, required: true, schema: { type: string, format: uuid } }
        - { name: limit, in: query, schema: { type: integer, default: 50, minimum: 1, maximum: 200 } }
        - { name: offset, in: query, schema: { type: integer, default: 0, minimum: 0 } }
        - { name: sort, in: query, schema: { type: string, enum: [title, added, year], default: title } }
        - { name: tag, in: query, description: "Filter to items carrying this tag (e.g. anime).", schema: { type: string } }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MediaItemPage"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /api/v1/libraries/{libraryId}/series:
    get:
      tags: [library]
      operationId: listSeries
      summary: Browse series in a library
      parameters:
        - { name: libraryId, in: path, required: true, schema: { type: string, format: uuid } }
        - { name: limit, in: query, schema: { type: integer, default: 50, minimum: 1, maximum: 200 } }
        - { name: offset, in: query, schema: { type: integer, default: 0, minimum: 0 } }
        - { name: sort, in: query, schema: { type: string, enum: [title, year], default: title } }
        - { name: tag, in: query, description: "Filter to series carrying this tag (e.g. anime).", schema: { type: string } }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SeriesPage"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /api/v1/series/{seriesId}:
    get:
      tags: [library]
      operationId: getSeries
      summary: Series detail with seasons and episodes
      parameters:
        - { name: seriesId, in: path, required: true, schema: { type: string, format: uuid } }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SeriesDetail"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
  /api/v1/items/{itemId}:
    get:
      tags: [library]
      operationId: getMediaItem
      summary: Media item detail with effective metadata
      parameters:
        - { name: itemId, in: path, required: true, schema: { type: string, format: uuid } }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MediaItemDetail"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
  /api/v1/items/{itemId}/stream:
    get:
      tags: [library]
      operationId: streamItem
      summary: Direct-play media stream with HTTP range support
      description: |
        Streams the item's media file with byte-range support (seeking). Auth is
        the per-device token via the bearer header OR a `token` query param,
        since an HTML5 `<video>` element cannot set Authorization.
      security: []
      parameters:
        - { name: itemId, in: path, required: true, schema: { type: string, format: uuid } }
        - { name: token, in: query, description: "Per-device token (alternative to the bearer header).", schema: { type: string } }
      responses:
        "200":
          description: Full media content
          content:
            application/octet-stream:
              schema: { type: string, format: binary }
        "206":
          description: Partial content (range request)
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
  /api/v1/items/{itemId}/subtitles:
    get:
      tags: [library]
      operationId: listSubtitles
      summary: Available subtitle tracks for an item
      description: |
        Lists subtitle tracks resolved through the priority chain: embedded text
        tracks (from the container) plus OpenSubtitles candidates when configured.
        Each track's `id` is fetched as WebVTT from the sibling endpoint.
      parameters:
        - { name: itemId, in: path, required: true, schema: { type: string, format: uuid } }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/SubtitleTrack"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
  /api/v1/items/{itemId}/subtitles/{trackId}:
    get:
      tags: [library]
      operationId: getSubtitle
      summary: A subtitle track as WebVTT
      description: |
        Produces (and disk-caches) the track as WebVTT. Auth is the per-device
        token via the bearer header OR a `token` query param, since an HTML5
        `<track>` element cannot set Authorization.
      security: []
      parameters:
        - { name: itemId, in: path, required: true, schema: { type: string, format: uuid } }
        - { name: trackId, in: path, required: true, description: "Track id from the list endpoint (e.g. `embedded:3`, `os:12345`).", schema: { type: string } }
        - { name: token, in: query, description: "Per-device token (alternative to the bearer header).", schema: { type: string } }
      responses:
        "200":
          description: WebVTT subtitle
          content:
            text/vtt:
              schema: { type: string }
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "502":
          description: Subtitle could not be produced (e.g. external fetch failed)
  /api/v1/items/{itemId}/playback:
    get:
      tags: [library]
      operationId: getPlaybackInfo
      summary: Direct-play capability decision for the item
      parameters:
        - { name: itemId, in: path, required: true, schema: { type: string, format: uuid } }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PlaybackInfo"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
  /api/v1/items/{itemId}/progress:
    get:
      tags: [library]
      operationId: getProgress
      summary: Current play-state (resume position) for the item
      parameters:
        - { name: itemId, in: path, required: true, schema: { type: string, format: uuid } }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PlayState"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
    put:
      tags: [library]
      operationId: reportProgress
      summary: Heartbeat — upsert the resume position for the current profile
      parameters:
        - { name: itemId, in: path, required: true, schema: { type: string, format: uuid } }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProgressUpdate"
      responses:
        "200":
          description: Updated play-state
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PlayState"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
  /api/v1/items/{itemId}/watched:
    post:
      tags: [library]
      operationId: setWatched
      summary: Mark an item watched / unwatched for the current profile
      parameters:
        - { name: itemId, in: path, required: true, schema: { type: string, format: uuid } }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WatchedUpdate"
      responses:
        "200":
          description: Updated play-state
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PlayState"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
  /api/v1/continue:
    get:
      tags: [library]
      operationId: listContinue
      summary: Continue-watching / on-deck items for the current profile
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ContinueItem"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /api/v1/recent:
    get:
      tags: [library]
      operationId: listRecent
      summary: Recently-added items (films + series) across the account
      description: >
        A unified "newly arrived" feed across every library: standalone films
        and series, newest first. A series' arrival time is the most recent of
        its episodes' added-at timestamps, so adding an episode resurfaces the
        series (deduped to a single card) rather than listing each episode.
        Each item's `kind` is "movie" or "series".
      parameters:
        - name: limit
          in: query
          required: false
          schema: { type: integer, default: 24, minimum: 1, maximum: 100 }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/MediaItemSummary"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /api/v1/sessions:
    get:
      tags: [library]
      operationId: listPlaybackSessions
      summary: Live playback sessions (who's watching what, where, now)
      description: >
        Active playback sessions, most-recently-active first. Admins see the
        whole account; a viewer sees only their own. A session that owns a
        transcode session is annotated with its encoder + method (reconciled
        with The Helm).
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/PlaybackSession"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /api/v1/beacon:
    get:
      tags: [library]
      operationId: beaconStream
      summary: SSE stream of the current user's live play-state events (Beacon)
      description: >
        Server-Sent Events stream of play_state changes for the authenticated
        user, powering cross-device resume/handoff. Authenticates via ?token=
        (an EventSource cannot set the Authorization header). Each message is
        `event: position` with a JSON data payload (see PlaybackSession-like
        fields). The client (EventSource) auto-reconnects; on reconnect it
        reconciles missed updates via a /continue or /progress fetch.
      parameters:
        - name: token
          in: query
          required: false
          schema: { type: string }
          description: Per-device bearer token (EventSource can't set headers).
      responses:
        "200":
          description: An SSE event stream
          content:
            text/event-stream:
              schema: { type: string }
        "401":
          $ref: "#/components/responses/Unauthorized"
  /api/v1/items/{itemId}/transcode:
    post:
      tags: [transcode]
      operationId: startTranscode
      summary: Start (or join) an HLS transcode session for an item
      description: |
        Starts a server-side transcode and returns the session plus its HLS
        playlist URL. Repeat requests for the same item + offset join the
        existing session instead of spawning a second ffmpeg. Returns 503 when
        the server is at its transcode-session limit (back-pressure).
      parameters:
        - { name: itemId, in: path, required: true, schema: { type: string, format: uuid } }
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TranscodeStartRequest"
      responses:
        "202":
          description: Session started or joined
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TranscodeSession"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "503":
          description: At transcode-session capacity
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /api/v1/transcode/sessions:
    get:
      tags: [transcode]
      operationId: listTranscodeSessions
      summary: Live transcode sessions for the current account (The Helm)
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TranscodeSession"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /api/v1/transcode/capabilities:
    get:
      tags: [transcode]
      operationId: getTranscodeCapabilities
      summary: Encoders available on this host and the selected one
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TranscodeCapabilities"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /api/v1/transcode/cache:
    get:
      tags: [transcode]
      operationId: getTranscodeCache
      summary: Transcode cache usage (Ballast)
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TranscodeCacheStats"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /api/v1/transcode/{sessionId}:
    delete:
      tags: [transcode]
      operationId: stopTranscode
      summary: Stop a transcode session and purge its output
      parameters:
        - { name: sessionId, in: path, required: true, schema: { type: string } }
      responses:
        "204":
          description: Stopped
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
  /api/v1/transcode/{sessionId}/{file}:
    get:
      tags: [transcode]
      operationId: getTranscodeFile
      summary: HLS artifact for a session (master/variant playlist, init, or segment)
      description: |
        Serves a session's HLS artifacts: the master playlist (`index.m3u8`),
        per-variant playlists (`stream_N.m3u8`), fMP4 init segments
        (`init_N.mp4`), and media segments (`stream_N_NNNNN.m4s`). The filename
        is allow-listed server-side. `index.m3u8` returns 503 while the session
        is still starting (no playlist written yet).
      parameters:
        - { name: sessionId, in: path, required: true, schema: { type: string } }
        - { name: file, in: path, required: true, schema: { type: string } }
      responses:
        "200":
          description: HLS playlist or segment
          content:
            application/vnd.apple.mpegurl:
              schema: { type: string }
            video/mp4:
              schema: { type: string, format: binary }
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "503":
          description: Session still starting (no playlist yet)
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: Per-device token issued by registerDevice.
  responses:
    Unauthorized:
      description: Missing or invalid credentials
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
...
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
  schemas:
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: string
    TranscodeStartRequest:
      type: object
      properties:
        startAt:
          type: number
          format: double
          description: Seek offset in seconds (default 0).
        hevc:
          type: boolean
          description: >
            Whether the client can play HEVC (H.265) in fMP4/MSE — detected via
            MediaSource.isTypeSupported. When true, an HEVC source is remuxed
            (copied) at native resolution including 4K instead of being
            re-encoded to H.264 1080p, and re-encodes of >1080p sources target
            HEVC. Defaults to false (H.264 only).
    TranscodeProgress:
      type: object
      required: [outTimeMs, speed, fps]
      properties:
        outTimeMs:
          type: integer
          format: int64
          description: Encoded timeline position in milliseconds.
        speed:
          type: number
          format: double
          description: Encode speed as a multiple of realtime (2.5 == 2.5x).
        fps:
          type: number
          format: double
    TranscodeSession:
      type: object
      required: [id, itemId, encoder, method, state, startAt, startedAt, playlistUrl, progress]
      properties:
        id:
          type: string
        itemId:
          type: string
          format: uuid
        encoder:
          type: string
        method:
          type: string
          enum: [remux, transcode]
          description: Whether this session copies codecs (remux) or re-encodes (transcode).
        state:
          type: string
          enum: [starting, running, complete, failed, stopped]
        startAt:
          type: number
          format: double
        startedAt:
          type: string
          format: date-time
        playlistUrl:
          type: string
          description: Relative URL of the HLS media playlist for this session.
        error:
          type: string
        progress:
          $ref: "#/components/schemas/TranscodeProgress"
    TranscodeCacheStats:
      type: object
      required: [totalBytes, budgetBytes, sessionDirs, liveDirs]
      properties:
        totalBytes: { type: integer, format: int64 }
        budgetBytes: { type: integer, format: int64 }
        sessionDirs: { type: integer }
        liveDirs: { type: integer }
    SubtitleTrack:
      type: object
      required: [id, source, language, label, forced, default]
      properties:
        id:
          type: string
          description: Track handle for the WebVTT endpoint (e.g. "embedded:3", "os:12345").
        source:
          type: string
          enum: [embedded, opensubtitles]
        language:
          type: string
          description: BCP-47 language code (e.g. "en"), or "und" when unknown.
        label:
          type: string
          description: Human-readable label for the subtitle picker.
        forced:
          type: boolean
        default:
          type: boolean
    TranscodeCapabilities:
      type: object
      required: [available, selected]
      properties:
        available:
          type: array
          items:
            type: string
          description: Encoder backends usable on this host (always includes "software").
        selected:
          type: string
          description: The encoder chosen by the configured preference order.
...
    PlaybackInfo:
      type: object
      required: [directPlay, method, container]
      properties:
        directPlay: { type: boolean }
        method:
          type: string
          enum: [direct, remux, transcode]
          description: Cheapest playable path — serve as-is, repackage (copy codecs), or re-encode.
        container: { type: string }
        videoCodec: { type: string, nullable: true }
        audioCodec: { type: string, nullable: true }
        reason: { type: string, nullable: true }
    PlayState:
      type: object
      required: [positionSeconds, watched]
      properties:
        positionSeconds: { type: number }
        durationSeconds: { type: number, nullable: true }
        watched: { type: boolean }
        updatedAt: { type: string, format: date-time, nullable: true }
    ProgressUpdate:
      type: object
      required: [positionSeconds]
      properties:
        positionSeconds: { type: number }
        durationSeconds: { type: number, nullable: true }
    WatchedUpdate:
      type: object
      required: [watched]
      properties:
        watched: { type: boolean }
    PlaybackSession:
      type: object
      required: [userId, deviceId, itemId, positionSeconds, state, startedAt, lastSeen]
      properties:
        userId: { type: string, format: uuid }
        deviceId: { type: string, format: uuid }
        itemId: { type: string, format: uuid }
        positionSeconds: { type: number, format: double }
        durationSeconds: { type: number, format: double, nullable: true }
        state: { type: string, description: "e.g. \"playing\"" }
        startedAt: { type: string, format: date-time }
        lastSeen: { type: string, format: date-time }
        encoder: { type: string, nullable: true, description: "Encoder of the owned transcode session, if any." }
        method: { type: string, nullable: true, description: "remux | transcode, if this session owns a transcode session." }
    ContinueItem:
      type: object
      required: [id, kind, title, positionSeconds, percent]
      properties:
        id: { type: string, format: uuid }
        kind: { type: string }
        title: { type: string }
        year: { type: integer, nullable: true }
        posterUrl: { type: string, nullable: true }
        positionSeconds: { type: number }
        durationSeconds: { type: number, nullable: true }
        percent: { type: number }
        seriesId: { type: string, format: uuid, nullable: true }
        seriesTitle: { type: string, nullable: true }
    Library:
      type: object
      required: [id, name, kind]
      properties:
        id: { type: string, format: uuid }
        name: { type: string }
        kind: { type: string }
    MediaItemSummary:
      type: object
      required: [id, kind, title, tags]
      properties:
        id: { type: string, format: uuid }
        kind: { type: string }
        title: { type: string }
        year: { type: integer, nullable: true }
        posterUrl: { type: string, nullable: true }
        tags: { type: array, items: { type: string } }
    MediaItemDetail:
      type: object
      required: [id, kind, title, filePath, reviewRequired, tags]
      properties:
        id: { type: string, format: uuid }
        kind: { type: string }
        title: { type: string }
        year: { type: integer, nullable: true }
        overview: { type: string, nullable: true }
        genres: { type: array, items: { type: string } }
        posterUrl: { type: string, nullable: true }
        durationSeconds: { type: number, nullable: true }
        container: { type: string, nullable: true }
        filePath: { type: string }
        reviewRequired: { type: boolean }
        tags: { type: array, items: { type: string } }
    MediaItemPage:
      type: object
      required: [items, total, limit, offset]
      properties:
        items: { type: array, items: { $ref: "#/components/schemas/MediaItemSummary" } }
        total: { type: integer }
        limit: { type: integer }
        offset: { type: integer }
    SeriesSummary:
      type: object
      required: [id, title, tags]
      properties:
        id: { type: string, format: uuid }
        title: { type: string }
        year: { type: integer, nullable: true }
        posterUrl: { type: string, nullable: true }
        tags: { type: array, items: { type: string } }
    SeriesPage:
      type: object
      required: [items, total, limit, offset]
      properties:
        items: { type: array, items: { $ref: "#/components/schemas/SeriesSummary" } }
        total: { type: integer }
        limit: { type: integer }
        offset: { type: integer }
    EpisodeSummary:
      type: object
      required: [id, episodeNumber]
      properties:
        id: { type: string, format: uuid }
        episodeNumber: { type: integer }
        title: { type: string, nullable: true }
        mediaItemId: { type: string, format: uuid, nullable: true }
        durationSeconds: { type: number, nullable: true }
        positionSeconds: { type: number, nullable: true, description: "Current profile's resume position." }
        watched: { type: boolean, nullable: true }
    SeasonSummary:
      type: object
      required: [id, seasonNumber, episodes]
      properties:
        id: { type: string, format: uuid }
        seasonNumber: { type: integer }
        title: { type: string, nullable: true }
        episodes: { type: array, items: { $ref: "#/components/schemas/EpisodeSummary" } }
    SeriesDetail:
      type: object
      required: [id, title, seasons, tags]
      properties:
        id: { type: string, format: uuid }
        title: { type: string }
        year: { type: integer, nullable: true }
        overview: { type: string, nullable: true }
        posterUrl: { type: string, nullable: true }
        seasons: { type: array, items: { $ref: "#/components/schemas/SeasonSummary" } }
        tags: { type: array, items: { type: string } }
security:
  - bearerAuth: []